Grab locale data before initializing app if not english
