+++ src/constants/locale.js
@@ -1,0 +1,17 @@
+export default {
+  en: {
+    value: 'en',
+    name: 'English',
+    localeFetch: () => import('react-intl/locale-data/en'),
+  },
+  fr: {
+    value: 'fr',
+    name: 'French',
+    localeFetch: () => import('react-intl/locale-data/fr'),
+  },
+  de: {
+    value: 'de',
+    name: 'German',
+    localeFetch: () => import('react-intl/locale-data/de'),
+  },
+};
